+++ qupath-gui-fx/src/main/java/qupath/lib/gui/panes/ProjectBrowser.java
@@ -39,14 +39,12 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.Random;
-import java.util.Set;
 import java.util.TreeMap;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
-
-import javax.imageio.ImageIO;
 
 import org.controlsfx.control.MasterDetailPane;
 import org.controlsfx.control.action.Action;
@@ -104,9 +102,7 @@ import qupath.lib.gui.tools.GuiTools;
 import qupath.lib.gui.tools.MenuTools;
 import qupath.lib.gui.tools.PaneTools;
 import qupath.lib.images.ImageData;
-import qupath.lib.images.servers.ImageServer;
-import qupath.lib.images.servers.ImageServerProvider;
-import qupath.lib.images.servers.ImageServers;
+import qupath.lib.images.servers.ImageServerMetadata;
 import qupath.lib.plugins.parameters.ParameterList;
 import qupath.lib.projects.Project;
 import qupath.lib.projects.ProjectIO;
@@ -138,7 +134,7 @@ public class ProjectBrowser implements ChangeListener<ImageData<BufferedImage>> 
 	private TreeView<Object> tree = new TreeView<>();
 
 	// Keep a record of servers we've requested - don't want to keep putting in requests if the server is unavailable
-	private Set<String> serversRequested = new HashSet<>();
+//	private Set<String> serversRequested = new HashSet<>();
 	
 	private StringProperty descriptionText = new SimpleStringProperty();
 	
@@ -938,21 +934,43 @@ public class ProjectBrowser implements ChangeListener<ImageData<BufferedImage>> 
 		Project<BufferedImage> project = qupath.getProject();
 		if (project == null) {
 			logger.error("Cannot set image name - project is null");
+			return false;
 		}
 		if (entry == null) {
 			logger.error("Cannot set image name - entry is null");
+			return false;
 		}
 		
 		String name = Dialogs.showInputDialog("Set Image Name", "Enter the new image name", entry.getImageName());
 		if (name == null)
 			return false;
+		
 		if (name.trim().isEmpty() || name.equals(entry.getImageName())) {
 			logger.warn("Cannot set image name to {} - will ignore", name);
+			return false;
 		}
 		
 		// Try to set the name
 		boolean changed = setProjectEntryImageName(entry, name);
 		if (changed) {
+			for (var viewer : qupath.getViewers()) {
+				var imageData = viewer.getImageData();
+				if (imageData == null)
+					continue;
+				var currentEntry = project.getEntry(imageData);
+				if (Objects.equals(entry, currentEntry)) {
+					var server = imageData.getServer();
+					if (!name.equals(server.getMetadata().getName())) {
+						// We update via the ImageData so that a property update is fired
+						var metadata2 = new ImageServerMetadata.Builder(server.getMetadata())
+								.name(name)
+								.build();
+						imageData.updateServerMetadata(metadata2);
+						// Bit of a cheat to force measurement table updates
+						imageData.getHierarchy().fireHierarchyChangedEvent(this);
+					}
+				}
+			}
 			tree.refresh();
 			qupath.refreshTitle();
 		}